Don't show sms option in fallback if sms disabled
Summary:
Original behavior is that developers can set "enableInitialSmsButton" setting to enable/disable SMS button on the initial screen, but in the fallback/resend screen, we always show the SMS option.

This changes the setting to "enableSms" which would determine whether we should use SMS at all, including removing the sms button from the fallback screen.

Reviewed By: ct2mak

Differential Revision: D16349445

fbshipit-source-id: d5e380a12ddc74c5f440fbff3e29cf715aae2415